dashboard

+++ containertransport/back_end/src/main/java/openerp/containertransport/service/DashboardService.java
@@ -3,5 +3,5 @@ package openerp.containertransport.service;
 import openerp.containertransport.dto.DashboardOrderByMonthRes;
 
 public interface DashboardService {
-    DashboardOrderByMonthRes getOrderByMonth(Long year);
+    DashboardOrderByMonthRes getOrderByMonth(int year);
 }
